Remember syscall arguments to probe buffer contents at sys_exit
This is needed to solve https://github.com/kinvolk/straceback/issues/2
where the buffer contents for read(2) are confusingly displayed because
they were probed at sys_enter where the kernel did not write the result
contents yet.
The remaining problem is that the parameter strings should be truncated
to the length of the returned buffer.
Ideally we would follow the behavior of strace and probe buffer contents
both at sys_enter and sys_exit depending on the syscall.