Use correct buffer lengths for argument strings
The pointer was just dereferenced and all 128 bytes copied which
were interpreted as C string. This caused the old buffer contents
to be shown for a EOF read(2) since the pointer was still valid.

Use either the maximum number of bytes (currently 128), or a
length argument (given through the arg index), or the return value
as argument length. Only copy this amount of data and use C.GoStringN
to dereference the pointer and interpret it as string.