Cleanup: Remove BPF debugs for each syscall and unused queue
We should not use printk in production since it generates a warning.
Also, it was printing too much in /sys/kernel/debug/tracing/trace_pipe
and makes it difficult to debug other BPF applications.

The queue feature was not used and was cleaned up here as well.

Supersedes: https://github.com/kinvolk/straceback/pull/6