Dump multiple times and concatenate dumps
Due to the limitations imposed by the backward perf ring buffers (they cannot
be read while written) traceloop was only able to dump the trace once.

This commit uses the new SwapAndDumpBackwards method that allows to perform
multiple dumps over the same perf ring buffer. The dumps are concatenated
before they are interpreted as strings because interpretation requires that
all events of a syscall are in one dump. The maximal length of the dumps is
specified by a constant for now. In interactive mode the buffer is swapped
four times per second. A marker for missed events when the ring buffer was
full would be needed to make clear that the concatenation can miss events.

Co-authored-by: Mauricio Vasquez <mauricio@kinvolk.io>
Signed-off-by: Kai Lüke <kai@kinvolk.io>