Fix the relation of syscall names to sys_enter_NAME
As found out in
https://github.com/kinvolk/traceloop/issues/12#issuecomment-563255909
the names of syscalls as known in /usr/include/asm/unistd_64.h
do not always relate to the names of the sys_enter_* kernel functions.

Create a helper function to relate the names of the functions to
the syscall names, and always only operate with the syscall names,
e.g., umount2 should be used instead of umount which comes from the
function name but does not refer to the the syscall name (yet, there
was once a syscall for umount but that is replaced by umount2).