guess offsets: increase theshold for threshold_nsproxy
On some kernels, traceloop fails with the following error message:
    "overflow while guessing uts namespace, bailing out"

It was noticed on:
- Ubuntu 18.04.4 LTS with Linux 4.15.0-91-generic
- Fedora 31 with Linux 5.5.5-200.fc31.x86_64

This happens when traceloop attempts to detect the offset of field
nsproxy in struct task_struct:
    offsetof(struct task_struct, nsproxy)

Before the patch, it was looking up to offset 2500 but on the
aforementioned kernels, the offset is further away.

In order to debug this, I have written the following code in BCC's
exectool:
https://github.com/kinvolk/bcc/commit/d9bc608f903978b0076c9abfd3eea1c05398d84e

and generated the docker image kinvolk/bcc:alban-uts-offset from it. It
can be executed with the following command:

docker run --rm -ti --privileged --net=host --pid=host \
    -v /usr/src:/usr/src \
    -v /lib/modules:/lib/modules \
    -v /sys/kernel/debug:/sys/kernel/debug \
    docker.io/kinvolk/bcc:alban-uts-offset \
    /usr/share/bcc/tools/execsnoop

to see new processes and then check the output of:

sudo cat /sys/kernel/debug/tracing/trace_pipe

to see something like:

    <...>-3025022 [002] .... 1149258.679821: 0: offset_nsproxy: 2784
    <...>-3025022 [002] .N.. 1149258.679824: 0: offset_utsns: 8
    <...>-3025022 [002] .N.. 1149258.679825: 0: offset_ns: 416
    <...>-3025022 [002] .N.. 1149258.679826: 0: offset_ino: 16

Here we see the offset is 2784, which is bigger than the previous
theshold of 2500. This patch increases the threshold.