Add a pseudo go module for bpf C code
The bpf C code is not really related in any way to any go code. It's
being compiled as a separate step with llvm and then the compiled
objects are put into a go code as a way of retrieving the objects from
the go code as binary blobs. But go tools still go into the directory
see the C code and see no go code importing "C", so they bail out with
a message like:

can't load package: package github.com/kinvolk/traceloop/bpf: C source
files not allowed when not using cgo or SWIG: straceback-guess-bpf.c
straceback-main-bpf.c straceback-tailcall-bpf.c

Since golang does not provide any way of ignoring directories for now,
make the bpf directory a submodule. Go tools usually work on a single
module, so the tool starting from the parent directory will not
traverse to the bpf directory, because the bpf directory is now a
different module.