Don't trace infrastructure containers
For every Kubernetes pod, there will be an additional container called
the "infrastructure container" holding the network namespace and running
a dummy process such as /pause or /usr/bin/pod (with cri-o).

When the BPF program detects a new container, it
- allocates a tracelet to that new container.
- sends a event of type CONTAINER_EVENT_TYPE_CREATE to the traceloop
  process in userspace via the perf ring buffer.

But at this stage, we don't know yet if it is an infrastructure
container. Traceloop will try to identify the new container with the
Kubernetes API by looking at the field Pod.Status.ContainerStatuses. If
no pods contain that new container ID for 15 seconds, we consider it is an
infrastructure container, and we remove it to free a slot for future
containers.