Handle pods with shared UTS namespaces
With Docker, each container of a pod has its own UTS namespace. With
cri-o, that might not be the case, depending on cri-o's version and its
manage_ns_lifecycle option.

Since traceloop currently separates traces by UTS namespaces, all
containers of a pod can be traced by the same tracelet.

This can include the infrastructure container (aka "pause" container or
/usr/bin/pod container). But the ContainerID of the infrastructure
container is not known by the Kubernetes API: the
Pod.Status.ContainerStatuses array does not include it. This confused
traceloop's procinformer when parsing /proc/*/ns/uts and finding the
infrastructure container ContainerID, and not finding it in the
Kubernetes API.

This patch remediates to that by verifying the ContainerID from the
procinformer with the Kubernetes API before accepting it. Because the
infrastructure container is started before the others, it might be that
the procinformer does not find the correct ContainerID immediately, so
this patch allows a few retries, but not more than 10 seconds.

Also, some renaming to make the code more clear:
- uid -> podUID
- UID -> PodUID (however the json marshalling stays the same to avoid
  incompatibility issues with Inspektor Gadget and Headlamp)
- info -> podInformerInfo or procInformerInfo (to avoid variable
  shadowing of different types)