Update scripts
The scripts now take a parameter being an output file. This will make
it possible to use them in the `go:generate` comments.

We also filter out an ID line from output params - it is likely to
differ from kernel to kernel and we don't care about it.

Other than that - minor cleanups, like using \t instead of literal
tabs.