Support building with Podman instead of Docker
The Makefile specifies the user of the Docker container to be the
current user ID to set the correct file permissions. This does not
work with Podman where the container's root user is mapped on the
current user and the current user ID inside the container has no
sufficient permissions to write the files.
Do not set the user ID when building with Podman and rely on
Podman to map the root user of the container to the current user.