Use read_str helper to copy NULL-terminalted syscall parameter buffers
The existing bpf_probe_read call had a failure in one corner case where
we didn't know the size of a NULL-terminated string and just copied as
much as fitted into the argument store. This failed when we reached the
end of a mapped memory area.
Use the bpf_probe_read_str helper to only copy as much as needed.
Later migrate to the other bpf_probe_read_user(_str) function because
we operate on userspace memory but since this requires Linux 5.5 it
is not done now.