Dereference pread64 buffers
The pread64 syscall is used in the benchmark script in the "contrib/"
folder.
We should try to generate the argument annotation with a script having
a heuristic for the data in
/sys/kernel/debug/tracing/events/syscalls/sys_enter_NAME/format
where "const" and "char *" are useful hints and then it will need to
fix some corner cases like dereferencing a read buffer at exit.