contrib: integrate with systemd units
The benchmark script in contrib/ already used ExecStartPre to
automatically set up systemd unit tracing. This was quite verbose and
and also lacked the "+" prefix to refer to the host filesystem and run
as root even if the unit processes are sandboxed.
Document how to modify systemd units so that they register with
traceloop. For convenience this is based on a traceloop service file
and new helper tool traceloopctl in the contrib folder. The
traceloopctl tool has a special systemd mode which follows a naming
scheme for a better integration but the traceloopctl can also be used
in the generic mode to interact with any trace.